+++ packages/dara-core/changelog.md
@@ -4,7 +4,8 @@ title: Changelog
 
 ## NEXT
 
--  Implement `Variable.init_override` static method to allow overriding how variables are initialized within a given context.
+-   Fixed an issue where in some cases a `DerivedDataVariable` would be invoked with an internal `PendingValue`
+-   Implement `Variable.init_override` static method to allow overriding how variables are initialized within a given context.
 
 ```python
 from dara.core import Variable